replace InputDescriptors class with array
