+++ src/Hyperledger.Aries/Features/SdJwt/Services/SdJwtVcHolderService/ISdJwtVcHolderService.cs
@@ -61,7 +61,7 @@ namespace Hyperledger.Aries.Features.SdJwt.Services.SdJwtVcHolderService
         /// </returns>
         Task<bool> DeleteAsync(IAgentContext context, string recordId);
         
-        Task<CredentialCandidates[]> GetCredentialCandidates(SdJwtRecord[] credentials, InputDescriptors inputDescriptors);
+        Task<CredentialCandidates[]> GetCredentialCandidates(SdJwtRecord[] credentials, InputDescriptor[] inputDescriptors);
         
         Task<string> CreateSdJwtPresentationFormatAsync(InputDescriptor inputDescriptors, string credentialId);
     }
