+++ test/Hyperledger.Aries.Tests/Hyperledger.Aries.Tests.csproj
@@ -40,7 +40,4 @@
   <ItemGroup>
     <None Remove="Features\Pex\Models\InputDescriptors.json" />
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Features\SdJwt\Models\" />
-  </ItemGroup>
 </Project>
